refactor: creating handy HTTP abstract objects
Its very important to make the code maintainable. We were parsing &
constructing the response std::basic_string<char> on the line just
before sending off the wire. But that is not scaleable and a very bad
practice. its too hard to manage things. Im building convinent HTTP
Abstraction objects for making things easy and in hoping for better
performance(I need to measure). Today, created the files and written a
decent progress on the implementation.